+++ src/app/Models/Club.model.ts
@@ -1,0 +1,6 @@
+export interface Club {
+  clubId: number;
+  name: string;
+  shortName: string;
+  imageId: string;
+}
