+++ src/app/Models/UserToken.model.ts
@@ -1,0 +1,7 @@
+export interface UserToken {
+  sub: string;       // ID/email
+  name?: string;     // Nombre completo
+  documentId?: string;
+  roles: string[];   // Roles del usuario
+  exp: number;       // Fecha expiración
+}
